remove helpers package
